requirements.txt
streamlit==0.58.0
pandas==1.0.1
numpy==1.18.1
matplotlib==3.1.3
python-slugify==4.0.0
tagging
